Commit: #1235 - Maximum Profit in Job Scheduling - Dynamic Programming Implementation
This commit includes the implementation of a dynamic programming solution for the LeetCode problem #1235, 'Maximum Profit in Job Scheduling'. The solution involves creating a vector to store the maximum profit for each job. For each job, we retrieve the job details and find the index of the latest non-conflicting job based on the job's start time. We then calculate the maximum profit for the current job by either including it or excluding it. The maximum profit that can be achieved is returned at the end.