The provided code is a SQL (Structured Query Language) query. SQL is a language used to interact with databases, and in this case, it's being used to retrieve data from a database.
The SELECT statement is used to select data from a database. The data returned is stored in a result table, called the result-set. Here, the SELECT statement is being used to select the name, population, and area columns from the database.

The FROM keyword indicates the table from which to retrieve the data. In this case, the data is being retrieved from the World table.

The WHERE clause is used to filter records. It's used to extract only those records that fulfill a specified condition. Here, it's being used to filter out countries that have an area of less than 3,000,000 square units (the units aren't specified in the query) and a population of less than 25,000,000 people. The OR keyword is used to filter records that satisfy either condition.

So, to sum it up, this SQL query is used to select the name, population, and area of all countries in the World table that have an area of 3,000,000 or more square units or a population of 25,000,000 or more.