Solve LeetCode #49 - Group Anagrams
The implemented solution uses a hash map to group anagrams together. For each string in the input vector, it creates a sorted version of the string to use as a key in the map. Anagrams will generate the same key when sorted. The original string is then moved into the vector associated with that key. Finally, all the vectors in the map (each representing a group of anagrams) are moved into a result vector which is then returned.