mplement majority element detection in an integer array"
This commit message succinctly describes the main change that has been made in the code. It mentions the key functionality that has been implemented, which is the detection of the majority element in an integer array.

Add SQL query to retrieve employees with bonus less than 1000 or no bonus"

This commit message clearly describes the main change that has been made in the code. It mentions the key functionality that has been implemented, which is the SQL query to retrieve employees with a bonus less than 1000 or no bonus at all.